+++ Craftsman/Builders/InfrastructureServiceRegistrationModifier.cs
@@ -33,8 +33,8 @@ using {servicesClassPath.ClassNamespace};";
             services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
                 .AddJwtBearer(options =>
                 {{
-                    options.Authority = Environment.GetEnvironmentVariable(""AUTH_AUTHORITY"");
-                    options.Audience = Environment.GetEnvironmentVariable(""AUTH_AUDIENCE"");
+                    options.Authority = EnvironmentService.Authority;
+                    options.Audience = EnvironmentService.Audience;
                     options.RequireHttpsMetadata = !env.IsDevelopment();
                 }});
         }}