+++ Craftsman/Builders/WebApiAppExtensionsBuilder.cs
@@ -13,22 +13,24 @@ public class WebApiAppExtensionsBuilder
         _utilities = utilities;
     }
 
-    public void CreateSwaggerWebApiAppExtension(string solutionDirectory, SwaggerConfig swaggerConfig, bool addJwtAuthentication, string projectBaseName)
+    public void CreateSwaggerWebApiAppExtension(string srcDirectory, SwaggerConfig swaggerConfig, bool addJwtAuthentication, string projectBaseName)
     {
-        var classPath = ClassPathHelper.WebApiApplicationExtensionsClassPath(solutionDirectory, $"SwaggerAppExtension.cs", projectBaseName);
-        var fileText = GetSwaggerAppExtensionText(classPath.ClassNamespace, solutionDirectory, swaggerConfig, addJwtAuthentication, projectBaseName);
+        var classPath = ClassPathHelper.WebApiApplicationExtensionsClassPath(srcDirectory, $"SwaggerAppExtension.cs", projectBaseName);
+        var fileText = GetSwaggerAppExtensionText(classPath.ClassNamespace, srcDirectory, swaggerConfig, addJwtAuthentication, projectBaseName);
         _utilities.CreateFile(classPath, fileText);
     }
 
-    public static string GetSwaggerAppExtensionText(string classNamespace, string solutionDirectory, SwaggerConfig swaggerConfig, bool addJwtAuthentication, string projectBaseName)
+    public static string GetSwaggerAppExtensionText(string classNamespace, string srcDirectory, SwaggerConfig swaggerConfig, bool addJwtAuthentication, string projectBaseName)
     {
-        var webApiClassPath = ClassPathHelper.WebApiMiddlewareClassPath(solutionDirectory, "", projectBaseName);
+        var webApiClassPath = ClassPathHelper.WebApiMiddlewareClassPath(srcDirectory, "", projectBaseName);
+        var envServiceClassPath = ClassPathHelper.WebApiServicesClassPath(srcDirectory, "", projectBaseName);
         return @$"namespace {classNamespace};
 
+using {webApiClassPath.ClassNamespace};
+using {envServiceClassPath.ClassNamespace};
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Swashbuckle.AspNetCore.SwaggerUI;
-using {webApiClassPath.ClassNamespace};
 
 public static class SwaggerAppExtension
 {{
@@ -39,8 +41,8 @@ public static class SwaggerAppExtension
     private static string GetSwaggerAppExtensionText(SwaggerConfig swaggerConfig, bool addJwtAuthentication)
     {
         var swaggerAuth = addJwtAuthentication ? $@"
-            config.OAuthClientId(Environment.GetEnvironmentVariable(""AUTH_CLIENT_ID""));
-            config.OAuthClientSecret(Environment.GetEnvironmentVariable(""AUTH_CLIENT_SECRET""));
+            config.OAuthClientId(EnvironmentService.ClientId);
+            config.OAuthClientSecret(EnvironmentService.ClientSecret);
             config.OAuthUsePkce();" : "";
 
         var swaggerText = $@"public static void UseSwaggerExtension(this IApplicationBuilder app)