+++ Craftsman/Domain/DbProvider.cs
@@ -16,6 +16,7 @@ public abstract class DbProvider : SmartEnum<DbProvider>
     public abstract string OTelSource();
     public abstract string TestingDbSetupMethod(string projectBaseName, bool isIntegrationTesting);
     public abstract string IntegrationTestConnectionStringSetup();
+    public abstract string ResetString();
     public abstract int Port();
     public abstract string DbConnectionStringCompose(string dbHostName, string dbName, string dbUser, string dbPassword);
     public abstract string DbConnectionString(string dbHostName, int? dbPort, string dbName, string dbUser, string dbPassword);
@@ -48,6 +49,26 @@ public abstract class DbProvider : SmartEnum<DbProvider>
 
         public override string IntegrationTestConnectionStringSetup() 
             => $@"Environment.SetEnvironmentVariable(""DB_CONNECTION_STRING"", _dbContainer.ConnectionString);";
+
+        public override string ResetString()
+        {
+            return $@"await using var connection = new NpgsqlConnection(EnvironmentService.DbConnectionString);
+        await connection.OpenAsync();
+        try
+        {{
+            var respawner = await Respawner.CreateAsync(connection, new RespawnerOptions
+            {{
+                TablesToIgnore = new Table[] {{ ""__EFMigrationsHistory"" }},
+                SchemasToExclude = new[] {{ ""information_schema"", ""pg_subscription"", ""pg_catalog"", ""pg_toast"" }},
+                DbAdapter = DbAdapter.Postgres
+            }});
+            await respawner.ResetAsync(connection);
+        }}
+        catch (InvalidOperationException e)
+        {{
+            throw new Exception($""There was an issue resetting your database state. You might need to add a migration to your project. You can add a migration with `dotnet ef migration add YourMigrationDescription`. More details on this error: {{e.Message}}"");
+        }}";
+        }
 
         public override int Port()
             => 5432;
@@ -93,7 +114,20 @@ public abstract class DbProvider : SmartEnum<DbProvider>
 
         public override string IntegrationTestConnectionStringSetup() 
             => $@"Environment.SetEnvironmentVariable(""DB_CONNECTION_STRING"", $""{{_dbContainer.ConnectionString}}TrustServerCertificate=true;"");";
-        
+        public override string ResetString()
+            => $@"try
+        {{
+            var respawner = await Respawner.CreateAsync(connection, new RespawnerOptions
+            {{
+                TablesToIgnore = new Table[] {{ ""__EFMigrationsHistory"" }},
+            }});
+            await respawner.ResetAsync(EnvironmentService.DbConnectionString);
+        }}
+        catch (InvalidOperationException e)
+        {{
+            throw new Exception($""There was an issue resetting your database state. You might need to add a migration to your project. You can add a migration with `dotnet ef migration add YourMigrationDescription`. More details on this error: {{e.Message}}"");
+        }}";
+
         public override int Port()
             => 1433;
         public override string DbConnectionStringCompose(string dbHostName, string dbName, string dbUser, string dbPassword)
@@ -113,6 +147,8 @@ public abstract class DbProvider : SmartEnum<DbProvider>
         public override string TestingDbSetupMethod(string projectBaseName, bool isIntegrationTesting)
             => throw new Exception(Response);
         public override string IntegrationTestConnectionStringSetup()
+            => throw new Exception(Response);
+        public override string ResetString() 
             => throw new Exception(Response);
         public override int Port()
             => throw new Exception(Response);